+++ src/main/groovy/netflix/nebula/dependency/recommender/DependencyRecommendationsPlugin.java
@@ -13,9 +13,6 @@ import org.gradle.api.logging.Logging;
 import org.gradle.api.plugins.ExtraPropertiesExtension;
 import org.gradle.api.plugins.JavaPlugin;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public class DependencyRecommendationsPlugin implements Plugin<Project> {
     private Logger logger = Logging.getLogger(DependencyRecommendationsPlugin.class);
 
@@ -34,14 +31,12 @@ public class DependencyRecommendationsPlugin implements Plugin<Project> {
                 project.getConfigurations().all(new Action<Configuration>() {
                     @Override
                     public void execute(final Configuration conf) {
-                        final List<String> firstOrderDepsWithoutVersions = new ArrayList<>();
-
+                        final RecommendationStrategyFactory rsFactory = new RecommendationStrategyFactory(project);
                         conf.getIncoming().beforeResolve(new Action<ResolvableDependencies>() {
                             @Override
                             public void execute(ResolvableDependencies resolvableDependencies) {
                                 for (Dependency dependency : resolvableDependencies.getDependencies()) {
-                                    if (dependency.getVersion() == null || dependency.getVersion().isEmpty())
-                                        firstOrderDepsWithoutVersions.add(dependency.getGroup() + ":" + dependency.getName());
+                                    rsFactory.getRecommendationStrategy().inspectDependency(dependency);
                                 }
                             }
                         });
@@ -50,7 +45,6 @@ public class DependencyRecommendationsPlugin implements Plugin<Project> {
                             @Override
                             public void execute(DependencyResolveDetails details) {
                                 ModuleVersionSelector requested = details.getRequested();
-                                String coord = requested.getGroup() + ":" + requested.getName();
 
                                 // don't interfere with the way forces trump everything
                                 for (ModuleVersionSelector force : conf.getResolutionStrategy().getForcedModules()) {
@@ -58,11 +52,9 @@ public class DependencyRecommendationsPlugin implements Plugin<Project> {
                                         return;
                                     }
                                 }
-
                                 String version = getRecommendedVersionRecursive(project, requested);
-                                if (version != null && firstOrderDepsWithoutVersions.contains(coord)) {
+                                if(rsFactory.getRecommendationStrategy().recommendVersion(details, version)) {
                                     logger.info("Recommending version " + version + " for dependency " + requested.getGroup() + ":" + requested.getName());
-                                    details.useVersion(version);
                                 }
                             }
                         });