+++ src/main/groovy/netflix/nebula/dependency/recommender/RecommendationStrategies.java
@@ -1,0 +1,16 @@
+package netflix.nebula.dependency.recommender;
+
+public enum RecommendationStrategies {
+    ConflictResolved(RecommendationsConflictResolvedStrategy.class),
+    OverrideTransitives(RecommendationsOverrideTransitivesStrategy.class);
+
+    private Class<? extends RecommendationStrategy> strategyClass;
+
+    RecommendationStrategies(Class<? extends RecommendationStrategy> strategyClass) {
+        this.strategyClass = strategyClass;
+    }
+
+    public Class<? extends RecommendationStrategy> getStrategyClass() {
+        return strategyClass;
+    }
+}